requirements.txt
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
joblib==1.3.2
statsmodels==0.14.0
scipy==1.11.3
plotly==5.18.0
python-dotenv==1.0.0 
openai
python-dotenv
